# docs/swagger.yaml
definitions:
  handler.requestProduct:
    properties:
      description:
        type: string
      expiration_rate:
        type: string
      freezing_rate:
        type: number
      height:
        type: number
      length:
        type: number
      net_weight:
        type: number
      product_code:
        type: string
      product_type_id:
        type: integer
      recommended_freezing_temperature:
        type: number
      width:
        type: number
    type: object
  web.Response:
    properties:
      code:
        type: integer
      data:
        type: any
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.ar/support
  description: This API Handle Mercado Fresco Sellers, Warehouse, Section, Products,
    Employees and Buyer
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: Mercado Fresco
  version: "1.0"
paths:
  /api/v1/products:
    get:
      consumes:
      - application/json
      description: get products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: We need token
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: store products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.requestProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: We need token
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.Response'
        "422":
          description: Missing some mandatory field
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store products
      tags:
      - Products
  /api/v1/products/{some_id}:
    delete:
      consumes:
      - application/json
      description: delete products by ID
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Some ID
        in: path
        name: some_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: We need ID
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: We need token
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete products by ID
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: list products by ID
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Some ID
        in: path
        name: some_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: We need ID
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: We need token
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/web.Response'
      summary: List products by ID
      tags:
      - Products
    patch:
      consumes:
      - application/json
      description: update products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Some ID
        in: path
        name: some_id
        required: true
        type: integer
      - description: Product to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.requestProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
        "400":
          description: We need ID
          schema:
            $ref: '#/definitions/web.Response'
        "401":
          description: We need token
          schema:
            $ref: '#/definitions/web.Response'
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/web.Response'
        "422":
          description: Missing some mandatory field
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update products by ID
      tags:
      - Products
swagger: "2.0"